+++ packages/server/sources/express/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @walkeros/server-source-express
 
+## 4.1.2
+
+### Patch Changes
+
+- @walkeros/collector@4.1.2
+- @walkeros/core@4.1.2
+
 ## 4.1.1
 
 ### Patch Changes