+++ packages/transformers/demo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/transformer-demo
 
+## 4.1.2
+
+### Patch Changes
+
+- @walkeros/core@4.1.2
+
 ## 4.1.1
 
 ### Patch Changes